Working Accessibility docker. Need some changes linked to geodecision. WIP.

+++ Accessibility/DockerContext/entrypoint.py
@@ -5,6 +5,8 @@ import shutil
 import re
 import datetime
 
+input_config = "/Input/config.json"
+
 subprocess.call(
         [
                 "conda", 
@@ -12,21 +14,25 @@ subprocess.call(
                 "-n", 
                 "geodecision", 
                 "python", 
-                "get_analysed_roofs.py", 
-                "/Input/config.json"
+                "accessibility_measures.py", 
+                input_config
                 ]
         )
 
-json_config=os.path.join('/Input', 'config.json')
-with open(json_config) as f: 
-   params = json.load(f)
-   output_dir = params["dir"]["output"]
-   target_output_dir = os.path.join("/Output", output_dir)
-
-if os.path.isdir(target_output_dir):
-    target_output_dir = target_output_dir + "_" + re.sub(
-            r"[-: ]", "_", str(
-                    datetime.datetime.now()
-            ).split(".")[0]
-    )
-shutil.copytree(output_dir, target_output_dir)
+if os.path.isfile(input_config):
+    with open(input_config) as f: 
+       params = json.load(f)
+       output_dir = params["output_folder"]
+       target_output_dir = os.path.join("/Output", output_dir)
+    
+    if os.path.isdir(target_output_dir):
+        target_output_dir = target_output_dir + "_" + re.sub(
+                r"[-: ]", "_", str(
+                        datetime.datetime.now()
+                ).split(".")[0]
+        )
+    shutil.copytree(output_dir, target_output_dir)
+else: 
+    raise FileNotFoundError(
+            "ERROR: Please check if {} exists".format(input_config)
+            )
